pyproject.toml
[project]
name = "mcp-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=2.14.3",
    "mysql-connector-python>=9.5.0",
    "pandas>=2.3.3",
]
